pyproject.toml
[tool.poetry]
name = "enderdragonkiller"
version = "0.1.0"
description = "A Minecraft bot to kill the enderdragon powered by LLMs"
authors = ["Zijian Zhang <futrime@outlook.com>"]
readme = "README.md"
packages = [{ include = "enderdragonkiller" }]

[tool.poetry.dependencies]
python = "^3.11"
javascript = "^1!1.0.3"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
